api/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-generativeai
langchain
langchain-google-genai
playwright
firebase-admin
httpx
python-dotenv
pytest
crawl4ai
json-repair
playwright-stealth
